Scriptable devbox

+++ devbox.sh
@@ -1,17 +1,28 @@
 #!/usr/bin/env bash
-set -e -u
+set -e -u -o pipefail
 
-TMP=$(mktemp)
+#
+if [ $# -lt 1 ]; then
+    echo "Usage: $0 <env> [command...]" >/dev/stderr
+    exit 1
+fi
+ROOT="$HOME/Documents/Systems/$1"
+shift
+
+# sudo closes all file descriptors except stdin, stdout, and stderr. Therefore
+# we must use this file on disk to send the script to the sandbox shell.
+TMP=$(mktemp -u)
+mkfifo "$TMP"
 cleanup() { rm "$TMP"; }
 trap cleanup EXIT
 
-ROOT="$HOME/Documents/Arch"
-cat <<EOF >"$TMP"
+# Begin writing the script before the shell is launched to avoid deadlock.
+cat <<EOF >$TMP &
 $(
-    for DIR in /dev /proc /sys $HOME; do
+    for DIR in /dev /proc /sys /lib/modules $HOME; do
         cat <<EOF2
-[ -d "$ROOT/$DIR" ] || mkdir -p "$ROOT/$DIR"
-mount --rbind "$DIR" "$ROOT/$DIR"
+[ -d "$ROOT$DIR" ] || mkdir -p "$ROOT$DIR"
+mount --rbind "$DIR" "$ROOT$DIR"
 EOF2
     done
 )
@@ -22,19 +33,21 @@ mkdir -p "$ROOT/run/dbus"
 mount --bind "/run/dbus" "$ROOT/run/dbus"
 # Also bind user sockets (necessary for graphical programs).
 if [ -n "${XDG_RUNTIME_DIR:-}" ]; then
-    mkdir -p "$ROOT/$XDG_RUNTIME_DIR"
-    mount --bind "$XDG_RUNTIME_DIR" "$ROOT/$XDG_RUNTIME_DIR"
+    mkdir -p "$ROOT$XDG_RUNTIME_DIR"
+    mount --bind "$XDG_RUNTIME_DIR" "$ROOT$XDG_RUNTIME_DIR"
 fi
-# The container will use the host's network interfaces but DNS config must be
-# forwarded too.
+# Also use the same DNS configuration as the host.
 [ -e "$ROOT/etc/resolv.conf" ] || touch "$ROOT/etc/resolv.conf"
 mount -o ro --bind "/etc/resolv.conf" "$ROOT/etc/resolv.conf"
-# Enter a new filesystem namespace.
+# Enter a new filesystem namespace and execute commands.
 hostname devbox
-exec chroot "$ROOT" /usr/bin/env bash -c "cd \\"$(pwd)\\" && bash -i"
+if [ \$# -gt 0 ]; then
+    exec unshare --root "$ROOT" --wd "$(pwd)" -- "\$@"
+else
+    exec unshare --root "$ROOT" --wd "$(pwd)" -- /usr/bin/env bash -i
+fi
 EOF
-
 # Enter a new mount namespace, which avoids so many problems that come with
 # binding the host filesystem to the container tree. Also enter a new time
 # sharing namespace so we can change the hostname.
-sudo --preserve-env -- unshare --mount --uts -- "$BASH" "$TMP"
+exec sudo --preserve-env -- unshare --mount --uts -- "$BASH" -- "$TMP" "$@"
